+++ lib/features/lock/set_pin/create_pin_update_pin.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bugfender/flutter_bugfender.dart';
 import 'package:pinput/pinput.dart';
 import 'package:msbridge/widgets/snakbar.dart';
 import 'package:provider/provider.dart';
-import 'package:msbridge/core/provider/app_pin_lock_provider.dart';
+import 'package:msbridge/core/provider/lock_app/app_pin_lock_provider.dart';
 
 class PinLockScreen extends StatefulWidget {
   final bool isCreating;
